pyproject.toml
[project]
name = "exam-modul7"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "argon2-cffi>=25.1.0",
    "django>=5.2.6",
    "django-filter>=25.1",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.28.0",
    "faker>=37.8.0",
    "pillow>=11.3.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
]
